deps: remove core_graphics and core_foundation
They are too large and not very ergonomic, therefore create a smaller,
more Rusty FFI for the Core Graphics and Core Foundation libraries.

Note that this is a first attempt and has some definite room for
improvement in further commits; notably:
- Removing the CF and CG prefixes to type names.
- Better separation of code into different files.
- More strict separation of code into bits (sys) and non-bits files.

As always, main.rs has been updated to replicate the function of previous,
but using the new lib features -- notably the Display struct.