axui: first rough draft for AXUI crate
main.rs show how to move a window programmatically using an AXUI element
ref. The correct window is determined based on the owner application's pid
and the position of the window. There may be better ways but this is the
best way I found from a cursory glance.

Code still needs a lot of cleaning-up.