ax_ui: observer: abstract AXUIObservers and notifications away from window
Create new struct that managers its own notifications. Rather than an
observer being a field of an ax_ui::Window, it is of main::Window. This
should make it cleaner when I add an observer to notify the main process
that a window has closed. Also just makes the code cleaner and shorter
anyway. (Still requires a fair amount of refactoring -- was the product of
a lot of headache and so is in a "this works finally" state).