main: poll for windows periodically to capture new windows
Since there is no way to detect new windows opening globally, have to just
periodically poll for all windows again. The AXUI library only provides a
way to detect when an application opens another window, using its pid.
Therefore it is impossible to detect a new window from an application that
has no windows currently on the screen, as far as I'm aware.

This also makes the previous commit relating to detecting window closure
(and all the nonsense with Rc and RefCells) completely pintless. It will
be ripped out in the next commit.