main: remove observers on Window structs
Rather than using Observers and callbacks, which adds considerable
complexity to the windows and leads to common segmentation faults due to
race conditions with setting up windows vs their observers, just move
windows to the position they "should" be according to the window manager,
every loop iteration. This produces the exact same effect, but adds a small
overhead; this is to be fixed in upcomming commits.